--- tests/neg-custom-args/convertible.check
@@ -1,21 +1,0 @@
--- [E120] Naming Error: tests/neg-custom-args/convertible.scala:15:6 ---------------------------------------------------
-15 |trait C[X] extends A[X]:  // error
-   |      ^
-   |      Name clash between defined and inherited member:
-   |      def f(x: X): Unit in trait A at line 10 and
-   |      override def f(x: into X): Unit in trait C at line 16
-   |      have the same type after erasure.
-   |
-   |      Consider adding a @targetName annotation to one of the conflicting definitions
-   |      for disambiguation.
--- [E120] Naming Error: tests/neg-custom-args/convertible.scala:18:6 ---------------------------------------------------
-18 |class D[X] extends B[X], C[X] // error
-   |      ^
-   |      Name clash between inherited members:
-   |      override def f(x: X): Unit in trait B at line 13 and
-   |      override def f(x: into X): Unit in trait C at line 16
-   |      have the same type after erasure.
-   |
-   |      Consider adding a @targetName annotation to one of the conflicting definitions
-   |      for disambiguation.
-there was 1 feature warning; re-run with -feature for details
