+++ tests/pos/into-sam.scala
@@ -1,0 +1,18 @@
+
+//> using options -feature -Xfatal-warnings
+
+import language.experimental.into
+
+class Text(val str: String)
+
+given Conversion[String, Text] = Text(_)
+object Test:
+  def f(x: Int)(y: into Text): Unit = ()
+  val _: Text => Unit = f(3)
+
+  trait ConvArg:
+    def apply(x: into Text): Unit
+
+  val x: ConvArg = f(3)(_)
+
+  x("abc")
